Add firebase configuration
For now I created 2 new apps on existing firebase project. If the clients would like to use their own firebase project this this files would need to be added to Github Secrets and injected before build. Or if we use the existing com.humhub.app and the our fcm service works as a proxy.